Fixed improperly formated SetTileState64 packets
There were 4 bytes of data in the packets that I did not see documented on Lifx's website. It appears to be a duration field which describes how long it takes the light to transition to the new color. I set this to zero, for an instant transition, but I may make it settable later on. should now give the proper color assigned near the top of the file.