Added Feature for controling the center tile, and a 12hour/24hour switch
I added a few new variables to allow you to choose between a 12 hour clock and a 24 hour clock.

I also added a colon character to the bitmaps, and added the option to choose what to do with the center tile. It can be ignored (leave it unchanged) set it to all , background color (default) all forground character, or display a colon character.