More enhancements
Changed the default Clock_Mode to 12 hour.
Cleaned up some of the extra whitespace.
Stopped it from printing the time every minute.
Added a timeout to the socket (default 5 seconds) so that the program will kill itself if the Tile is not found. This will properly close the socket and allow it to be re-opened immediately. I did this because I found that sometimes the program would not see a response from the Tile during the discovery stage. This doesn't solve the problem, just handles it a little better.